UI: Add Search to Hotkeys settings menu
This commit adds a search bar to the top of the hotkeys settings menu,
with it set to look for matches anywhere in the string label for that
hotkey.

It currently has issues when clearing a search, since setVisible(true)
or show() on a widget will recompute the styles, making it a slow
operation. There doesn't seem to be a way around this without
implementing pagination or some other means of displaying the data
first.