cmake: Remove local files for checking threading support
These local copies of CheckForPthreads.c and FindThreads.cmake override
the ones included with CMake, which Qt6 expects to use. These
local/custom versions seem to be based on old copies from CMake from
late 2014, and they can probably be safely removed in favor of the
built-in ones that CMake uses.