libobs/media-io: Preserve frame rate mode on remux
When remuxing from MKV to MP4/MOV, FFmpeg adjusts the timescale for any
video tracks/streams. This adjustment appears to cause some editors and
media info tools to detect and treat the file as Variable Frame Rate
instead of Constant Frame Rate. To prevent this timescale adjustment, we
can force the video track's to use a specific integer timescale. If we
use the input file's video track frame rate as the output file's video
track timescale, the file is detected as Constant Frame Rate.