obs-ffmpeg: Fix fractional framerates reading as VFR
When recording with NVENC via FFmpeg with B-frames enabled, the DTS
negative adjustment in FFmpeg relies on the framerate denominator and
ticks_per_frame. Not setting either value results in both being inferred
by FFmpeg, with ticks_per_frame defaulting to 1. When this happens, the
DTS negative adjustment value becomes -2, so the timestamp 0 becomes -2,
and 1001 becomes 999, instead of -2002 and -1001, respectively. Set
framerate and ticks_per_frame explicitly to handle fractional
framerates.