obs-ffmpeg: Use NVENC compatibility hack if non-AV1 and old drivers
If encoding in non-AV1 and using drivers older than the configured
SDK/API requires, fall back to the compatibility version of the API
(11.1).

An example scenario would be encoding in H.264 on driver version 527.37.
SDK 12.1 requires driver version 531.61 on Windows. Currently, in this
scenario, OBS will fall back to the FFmpeg encoder path, which will do
its own driver check, which will fail on this driver version.

With this change, this scenario will instead leverage the existing
compatibility hack to fall back to the compatibility version of the API
without falling back to FFmpeg.

It should be noted that while neat, the compatibility hack is due to be
removed in the next major release of OBS.